+++ app/src/main/java/com/ar/bootcampar/model/utilities/ICursorWrapper.java
@@ -1,21 +1,13 @@
-package com.ar.bootcampar.model;
+package com.ar.bootcampar.model.utilities;
 
 public interface ICursorWrapper {
     void close();
-
     int getCount();
-
     boolean moveToFirst();
-
     int getColumnIndex(String columnName);
-
     long getLong(int columnIndex);
-
     int getInt(int columnIndex);
-
     String getString(int columnIndex);
-
     boolean moveToNext();
-
     boolean isAfterLast();
 }
