+++ app/src/main/java/com/ar/bootcampar/model/utilities/ISQLiteDatabaseWrapper.java
@@ -1,26 +1,18 @@
-package com.ar.bootcampar.model;
+package com.ar.bootcampar.model.utilities;
 
-import android.content.ContentValues;
-import android.database.Cursor;
+import com.ar.bootcampar.model.utilities.IContentValuesWrapper;
+import com.ar.bootcampar.model.utilities.ICursorWrapper;
 
 public interface ISQLiteDatabaseWrapper {
     void close();
-
     ICursorWrapper query(String table, String[] columns, String selection,
-                 String[] selectionArgs, String groupBy, String having,
-                 String orderBy);
-
+                         String[] selectionArgs, String groupBy, String having,
+                         String orderBy);
     void endTransaction();
-
     void setTransactionSuccessful();
-
     void beginTransaction();
-
     long insert(String table, String nullColumnHack, IContentValuesWrapper values);
-
     int delete(String table, String whereClause, String[] whereArgs);
-
     int update(String table, IContentValuesWrapper values, String whereClause, String[] whereArgs);
-
     void execSQL(String sql);
 }
