+++ app/src/test/java/com/ar/bootcampar/ModificarGrupoDebe.java
@@ -10,7 +10,7 @@ import static org.junit.Assert.assertTrue;
 
 import com.ar.bootcampar.model.Database;
 import com.ar.bootcampar.model.Grupo;
-import com.ar.bootcampar.model.ISQLiteDatabaseWrapper;
+import com.ar.bootcampar.model.utilities.ISQLiteDatabaseWrapper;
 import com.ar.bootcampar.support.SqliteDatabaseWrapperSpy;
 import com.ar.bootcampar.support.TestableDatabase;
