+++ app/src/test/java/com/ar/bootcampar/support/TestableDatabase.java
@@ -1,9 +1,8 @@
 package com.ar.bootcampar.support;
 
 import com.ar.bootcampar.model.Database;
-import com.ar.bootcampar.model.IContentValuesWrapper;
-import com.ar.bootcampar.model.ISQLiteDatabaseWrapper;
-import com.ar.bootcampar.support.ContentValuesSpy;
+import com.ar.bootcampar.model.utilities.IContentValuesWrapper;
+import com.ar.bootcampar.model.utilities.ISQLiteDatabaseWrapper;
 
 public class TestableDatabase extends Database {
     private ISQLiteDatabaseWrapper databaseSpy;
